moved getMetadata to RatingReview parent component so rating summary is refreshed upon sort option change
