+++ client/src/components/RR/RatingReview.jsx
@@ -1,24 +1,39 @@
 import React from 'react';
+import axios from 'axios';
 import PropTypes from 'prop-types';
 import ReviewList from './Review/ReviewList';
 import RatingSummary from './Rating/RatingSummary';
+
+const ratingAverageAndCount = (ratingObj) => {
+  let sum = 0;
+  let count = 0;
+  const keys = Object.keys(ratingObj);
+  const values = Object.values(ratingObj);
+  for (let i = 0; i < keys.length; i += 1) {
+    sum += Number(keys[i]) * Number(values[i]);
+    count += Number(values[i]);
+  }
+  return { avg: (sum / count).toFixed(1), count };
+};
 
 class RatingReview extends React.Component {
   constructor(props) {
     super(props);
     this.state = {
-      sortOption: 'relevant',
       filter: [],
+      metaData: {},
     };
     this.handleRatingFilterClick = this.handleRatingFilterClick.bind(this);
     this.handleRemoveFilterClick = this.handleRemoveFilterClick.bind(this);
     this.handleSortSelection = this.handleSortSelection.bind(this);
   }
 
-  handleSortSelection(sortOption) {
-    this.setState({
-      sortOption,
-    });
+  componentDidMount() {
+    this.getReviewMetadata();
+  }
+
+  handleSortSelection() {
+    this.getReviewMetadata();
   }
 
   handleRatingFilterClick(star) {
@@ -41,6 +56,48 @@ class RatingReview extends React.Component {
     });
   }
 
+  getReviewMetadata() {
+    // const { selected } = this.props;
+    // Test for product with no review
+    // const selected = 40346;
+
+    // Test for product with different reviews
+    const selected = 40345;
+    // const selected = 40347;
+    axios.get('/reviews/meta/', { params: { product_id: selected } })
+      .then((response) => {
+        const { ratings, recommended, characteristics } = response.data;
+        if (!Object.keys(ratings).length) {
+          this.setState({
+            metaData: {
+              characteristics,
+              ratings,
+              avgRating: null,
+              ratingCount: 0,
+              recRate: null,
+            },
+          });
+        } else {
+          const ratingCalc = ratingAverageAndCount(ratings);
+          const rec = Number(recommended.true);
+          const notRec = Number(recommended.false);
+          const recRate = Math.round((rec * 100) / (rec + notRec));
+          this.setState({
+            metaData: {
+              characteristics,
+              ratings,
+              avgRating: Number(ratingCalc.avg),
+              ratingCount: ratingCalc.count,
+              recRate,
+            },
+          });
+        }
+      })
+      .catch((err) => {
+        console.log('Error getting review metadata: ', err);
+      });
+  }
+
   render() {
     // const { selected } = this.props;
     // Test for product with no review
@@ -49,13 +106,13 @@ class RatingReview extends React.Component {
     // Test for product with different reviews
     const selected = 40345;
     // const selected = 40347;
-    const { filter, sortOption } = this.state;
+    const { filter, metaData } = this.state;
     return (
       <div>
         <h2>Ratings and Reviews</h2>
         <div className="container">
           <RatingSummary
-            selected={selected}
+            metaData={metaData}
             handleRatingFilterClick={this.handleRatingFilterClick}
             handleRemoveFilterClick={this.handleRemoveFilterClick}
             filter={filter}
@@ -63,7 +120,6 @@ class RatingReview extends React.Component {
           <ReviewList
             selected={selected}
             filter={filter}
-            sortOption={sortOption}
             handleSortSelection={this.handleSortSelection}
           />
         </div>
@@ -73,12 +129,10 @@ class RatingReview extends React.Component {
 }
 
 RatingReview.propTypes = {
-  sortOption: PropTypes.string,
   selected: PropTypes.number,
 };
 
 RatingReview.defaultProps = {
-  sortOption: 'relevant',
   selected: 40344,
 };
